% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter.R
\name{buatfilter}
\alias{buatfilter}
\alias{create_filter}
\title{Membuat filter yang diinginkan}
\usage{
buatfilter(...)
create_filter(...)
}
\arguments{
\item{...}{Daftar filter yang diinginkan (dapat dibuat sebanyak mungkin).}
}
\value{
Filter yang dapat digunakan untuk memfilter data dalam fungsi \code{\link{penanda}}.
}
\description{
Membuat filter yang diinginkan
}
\examples{
filter1 = buatfilter(NAMA_PROV == "ACEH", KLASIFIKASI == "1", JenisKegiatan != 2)
filter2 = buatfilter(NAMA_PROV == "ACEH" | NAMA_PROV == "RIAU", KLASIFIKASI == "2", JenisKegiatan \%in\% c(1,2,3))
}
