% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_sample.R
\name{mutate_sample}
\alias{mutate_sample}
\title{Mengubah nilai atribut-atribut sampel}
\usage{
mutate_sample(x, sample_flag, ...)
}
\arguments{
\item{x}{Dataset yang digunakan.}

\item{sample_flag}{Flag dari sampel imputasi yang ingin diubah atribut-atributnya.}

\item{...}{Daftar mapping atribut dan nilai baru yang diinginkan (dapat dibuat sebanyak mungkin).}
}
\value{
Data yang telah diubah nilai atribut-atributnya.
}
\description{
Mengubah nilai atribut-atribut sampel dengan flag tertentu
}
\examples{
survei_dummy = mutate_sample(x = survei_dummy, sample_flag = "aceh_desa", status6 = NA, kategori = 1)
}
