% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filter.R
\name{create_filter}
\alias{create_filter}
\title{Membuat filter yang diinginkan.}
\usage{
create_filter(...)
}
\arguments{
\item{...}{Daftar filter yang diinginkan.}
}
\value{
Filter yang dapat digunakan untuk memfilter data dalam fungsi \code{\link{imputation_sample}}.
}
\description{
Membuat filter yang diinginkan.
}
\examples{
filter_list = create_filter(mpg > 15, gear == 4, between(hp, 100, 110))
}
