% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_sample.R
\name{imputation_sample}
\alias{imputation_sample}
\title{Memilih sampel imputasi}
\usage{
imputation_sample(x, filters, weight_aggregate, weight_col, iter = 1,
  flag = 1)
}
\arguments{
\item{x}{Dataset yang digunakan.}

\item{filters}{Filter yang telah dibuat dengan fungsi \code{\link{create_filter}}.}

\item{weight_aggregate}{Besaran agregat weight dari sampel terpilih yang diinginkan.}

\item{weight_col}{Kolom yang digunakan sebagai weight dalam pemilihan sampel apabila tersedia.}

\item{iter}{Jumlah iterasi pengacakan dan pengambilan sampel yang diinginkan (semakin tinggi maka weight akan semakin sesuai).}

\item{flag}{Identitas dari sampel yang dihasilkan.}
}
\value{
Filter yang dapat digunakan untuk memfilter data dalam fungsi \code{\link{imputation_sample}}.
}
\description{
Memilih sampel imputasi dari filter yang telah dibuat dengan total weight tertentu
}
\examples{
# Membuat filter berbeda
filter_1 = create_filter(NAMA_PROV == "ACEH", KLASIFIKASI == 1)
filter_2 = create_filter(NAMA_PROV == "RIAU" | NAMA_PROV == "SUMATERA BARAT", KLASIFIKASI == 2)
imputation_sample(x = sakernas_dummy, filters = filter_1, weight_aggregate = 10000, weight_col = Weight_R, flag = "aceh_1")
imputation_sample(x = sakernas_dummy, filters = filter_2, weight_aggregate = 5000, weight_col = Weight_R, flag = "riau_sumbar_2")

# Membandingkan hasil sampel dengan jumlah iterasi berbeda
my_filter = create_filter(NAMA_PROV == "SUMATERA BARAT", KLASIFIKASI == 2)
imputation_sample(x = sakernas_dummy, filters = my_filter, weight_aggregate = 73955, weight_col = Weight_R, iter = 1)
imputation_sample(x = sakernas_dummy, filters = my_filter, weight_aggregate = 73955, weight_col = Weight_R, iter = 100)
}
