% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation_sample.R
\name{imputation_sample}
\alias{imputation_sample}
\title{Memilih sampel imputasi dari filter yang telah dibuat.}
\usage{
imputation_sample(x, filters, size, weight_col = NULL, flag = 1)
}
\arguments{
\item{x}{Dataset yang digunakan.}

\item{filters}{Filter yang telah dibuat dengan fungsi \code{\link{create_filter}}.}

\item{size}{Besaran sampel yang diinginkan.}

\item{weight_col}{Kolom yang digunakan sebagai weight dalam pemilihan sampel apabila tersedia.}

\item{flag}{Identitas dari sampel yang dihasilkan}
}
\value{
Filter yang dapat digunakan untuk memfilter data dalam fungsi \code{\link{imputation_sample}}.
}
\description{
Memilih sampel imputasi dari filter yang telah dibuat.
}
\examples{
filter_list = create_filter(mpg > 15, gear == 4, between(hp, 100, 110))
imputation_sample(x = mtcars, filters = filter_list, size = 4)
imputation_sample(x = mtcars, filters = filter_list, size = 4, weight_col = disp, flag = "sampelku")
}
