% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_sample.R
\name{imputasi}
\alias{imputasi}
\alias{mutate_sample}
\title{Mengubah nilai atribut-atribut sampel}
\usage{
imputasi(d, flag, ...)
mutate_sample(x, sample_flag, ...)
}
\arguments{
\item{d}{Dataset yang digunakan.}

\item{flag}{Target baris data yang ingin diubah atribut-atributnya.}

\item{...}{Daftar mapping atribut dan nilai baru yang diinginkan (dapat dibuat sebanyak mungkin).}
}
\value{
Data yang telah diubah nilai atribut-atributnya.
}
\description{
Mengubah nilai atribut-atribut sampel dengan flag tertentu
}
\examples{
survei_dummy = imputasi(d = survei_dummy, flag = "aceh_desa", status6 = NA, kategori = 1)
}
